MomentPlayerStunt stages its World_Signature take by the ZERO-extended name hash: the signature-jump camera no longer asserts and faults at address 8 (FX-DIRECTOR)
The first live run with the moment tick on died on a jump
(scratch/bugtest/runs/fxdirector_moment_tick/20260924_180957): jump 4030 posted OnJumpStart cut 305,
MomentPlayerStunt fired "lpIceTake != NULL" (cpp:103), and ICETakeData_GetGuid read address 8.

The take id is the CRC32 of "World_Signature_<cut>". CgsResource::ID::HashString @0x828D84A8 returns
it ZERO-extended (`nor r11, r9, r9 ; clrldi r3, r11, 0x20` at 0x828D854C), and
MomentPlayerStunt::Update @0x82272750 hands that register straight to
DirectorResourceManager::GetKeyAnim (0x82272B04 `mr r4, r3`). The PC widened HashString's s32 return
straight to u64, which SIGN-extends every CRC with its top bit set. "World_Signature_305" is 0x9CA2368A,
so the lookup asked for 0xFFFFFFFF9CA2368A. build/game/CAMERAS.BUNDLE stores the take's key as the
zero-extended little-endian u64 (one hit) and has no sign-extended form (zero hits). The other
HashString widenings in the tree already go through u32 (BrnICEMoviePlayer, CgsGuiResourceModule).
MomentPlayerStunt_StageTakeReference now does the same. The helper serves both staging sites: the
allocation at 0x82272AE4 and the mid-flight take swap at 0x82272D2C.

Test: tests/run_fxdirector_stunt_take.py scores 7/7. It extracts the production helper and runs it with
the real HashString and SPrintf against a one-take dictionary keyed like the bundle:
* the 305 take is found with id 0x000000009CA2368A;
* a CRC with the top bit clear resolves either way;
* a missing take still raises the console's assert.
On the old body (--rev HEAD) it scores 4/7. Faithfulness gate PASS.
Live: tests/FxDirectorStuntJumpLive.ps1 seats the car on that run's approach. It ran with the tick on
through the lane's temporary, never-committed opt-in, and was GREEN 8/8
(scratch/bugtest/runs/fxdirector_stunt_jump/20260924_182507):
* `jump region latched (id=4030)`;
* `OnJumpStart posted cut=305 type=1`;
* then `[ice-prepare] guid 572230 takeData 1 len 4.025853`: the World_Signature_305 take was found and
  prepared;
* 0 asserts, 0 exceptions.

+++ src/GameSource/Director/MomentController/Moments/BrnMomentPlayerStunt.cpp
@@ -92,9 +92,17 @@ static void MomentPlayerStunt_StageTakeReference(const void* lSharedInfo,
 
     char lacTakeName[64];
     CgsCore::SPrintf(lacTakeName, 64, "World_Signature_%i", liTakeNumber);
+    // The id is the name's CRC32 ZERO-extended to 64 bits: HashString @0x828D84A8 returns through
+    // `nor r11, r9, r9 ; clrldi r3, r11, 0x20` (0x828D854C), and GetKeyAnim takes that register as
+    // the id (0x82272B04 `mr r4, r3`).
+    // ⭐ 2026-09-24 (FX-DIRECTOR): this used to widen the s32 return straight to u64, which
+    // SIGN-extends every name whose CRC has its top bit set -- "World_Signature_305" is
+    // 0x9CA2368A -- so the dictionary lookup missed, the "lpIceTake != NULL" assert fired and
+    // the guid read below faulted at address 8 (scratch/bugtest/runs/fxdirector_moment_tick/
+    // 20260924_180957, the first live run with the moment tick on).
     CgsResource::ID lTakeId;
-    lTakeId.SetHash(static_cast<u64>(
-        CgsResource::ID::HashString(reinterpret_cast<const u8*>(lacTakeName))));
+    lTakeId.SetHash(static_cast<u64>(static_cast<u32>(
+        CgsResource::ID::HashString(reinterpret_cast<const u8*>(lacTakeName)))));
     ICE::ICETakeData* lpIceTake =
         MomentSharedInfo_GetDirectorResourceManager(lSharedInfo)->GetKeyAnim(lTakeId);
     // (One assert on the allocation path, one on the take swap -- both non-gating.)
